Added JWT-Extended, jwt-claims, jwt-optional

+++ resources/item.py
@@ -1,5 +1,6 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import (
+    jwt_required, get_jwt_claims, jwt_optional, get_jwt_identity)
 
 from models.item import ItemModel
 
@@ -18,14 +19,14 @@ class Item(Resource):
         help="Store ID is required"
     )
 
-    @jwt_required()
+    @jwt_required
     def get(self, iname):
         item = ItemModel.find_by_name(iname)
         if item:
             return item.json()
         return {"message": "Item Not Found!"}, 404
 
-    @jwt_required()
+    @jwt_required
     def post(self, iname):
         if ItemModel.find_by_name(iname):
             return {"message": "Item already exist!"}, 400
@@ -37,7 +38,7 @@ class Item(Resource):
 
         return {"message": "Item created!", "item": item.json()}, 201
 
-    @jwt_required()
+    @jwt_required
     def put(self, iname):
         item = ItemModel.find_by_name(iname)
         data = Item.parser.parse_args()
@@ -50,8 +51,11 @@ class Item(Resource):
 
         return {"message": "Item Updated!", "item": item.json()} 
         
-    @jwt_required()
+    @jwt_required
     def delete(self, iname):
+        claims = get_jwt_claims()
+        if not claims["is_admin"]:
+            return {"message": "Admin privileges required!"}, 401
         item = ItemModel.find_by_name(iname)
         if item:
             item.delete()
@@ -61,6 +65,12 @@ class Item(Resource):
 
 class ItemList(Resource):
 
-    @jwt_required()
+    @jwt_optional
     def get(self):
-        return {"items": [item.json() for item in ItemModel.find_all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {"items": items}
+        return {"items": [item["item"] for item in items],
+            "message": "For more details kindly login" 
+        }
